+++ src/routes/v1/proposals.ts
@@ -14,12 +14,15 @@ import {
   rejectNegotiation,
 } from '../../controllers/NegotiationController.js';
 import { CreateNegotiationSchema } from '../../schemas/requests/negotiation.request.js';
+import { authorizeApprovedWorker, authorizeWorker } from 'src/middlewares/workerMiddleware.js';
+import { authorizeClient } from 'src/middlewares/clientMiddleware.js';
 
 const proposalsRouter = Router({ mergeParams: true });
 
 proposalsRouter.post(
   '/',
   isActive,
+  authorizeApprovedWorker,
   validateParams(OrderIdParamsSchema),
   validateBody(CreateProposalSchema),
   proposalController.submit
@@ -28,22 +31,33 @@ proposalsRouter.post(
 proposalsRouter.get(
   '/',
   isActive,
+  authorizeClient,
   validateParams(OrderIdParamsSchema),
   proposalController.list
 );
 
 proposalsRouter.get(
+  '/mine',
+  isActive,
+  authorizeWorker,
+  validateParams(OrderProposalParamsSchema),
+  proposalController.getMine
+);
+
+proposalsRouter.get(
   '/:proposalId',
   isActive,
+  authorizeClient,
+  authorizeWorker,
   validateParams(OrderProposalParamsSchema),
   proposalController.getById
 );
 
-
-
 proposalsRouter.get(
   '/:proposalId/negotiations',
   isActive,
+  authorizeClient,
+  authorizeWorker,
   validateParams(OrderProposalParamsSchema),
   getNegotiations
 );
@@ -51,6 +65,8 @@ proposalsRouter.get(
 proposalsRouter.post(
   '/:proposalId/negotiations',
   isActive,
+  authorizeClient,
+  authorizeWorker,
   validateParams(OrderProposalParamsSchema),
   validateBody(CreateNegotiationSchema),
   createNegotiation
@@ -59,6 +75,8 @@ proposalsRouter.post(
 proposalsRouter.post(
   '/:proposalId/negotiations/accept',
   isActive,
+  authorizeClient,
+  authorizeWorker,
   validateParams(OrderProposalParamsSchema),
   acceptNegotiation
 );
@@ -66,6 +84,8 @@ proposalsRouter.post(
 proposalsRouter.post(
   '/:proposalId/negotiations/reject',
   isActive,
+  authorizeClient,
+  authorizeWorker,
   validateParams(OrderProposalParamsSchema),
   rejectNegotiation
 );